Network.PagerDuty.Integration

Description

The PagerDuty event integration API is how you would add PagerDuty's advanced alerting functionality to any system that can make an HTTP API call. You can now add phone, SMS and email alerting to your monitoring tools, ticketing systems and custom software.

The API was designed to allow you to easily integrate a monitoring system with a Service in PagerDuty. Monitoring systems generally send out events when problems are detected and when these problems have been resolved (fixed). Some more advanced systems also understand the concept of acknowledgements: problems can be acknowledged by an engineer to signal he or she is working on fixing the issue.

Since monitoring systems emit events, the API is based around accepting events. Incoming events (sent via the API) are routed to a PagerDuty service and processed. They may result in a new incident being created, or an existing incident being acknowledged or resolved.

The same event-based API can also be used to integrate a PagerDuty service with ticketing systems and various other software tools.

See: http://developer.pagerduty.com/documentation/integration/events

Events

submit :: (MonadIO m, Event a) => Manager -> a -> m (Either Error Response)

Send an event to the integration API.

submitWith :: (MonadIO m, Event a) => Manager -> Logger -> a -> m (Either Error Response)

See: submit

Trigger

data Trigger

trigger

Arguments

:: ServiceKey

-> Text
tDescription

-> Trigger

Your monitoring tools should send PagerDuty a trigger event to report a new or ongoing problem. When PagerDuty receives a trigger event, it will either open a new incident, or add a new trigger log entry to an existing incident, depending on the provided incident_key.

tDescription :: Lens' Trigger Text

A short description of the problem that led to this trigger. This field (or a truncated version) will be used when generating phone calls, SMS messages and alert emails. It will also appear on the incidents tables in the PagerDuty UI. The maximum length is 1024 characters.

tIncidentKey :: Lens' Trigger (Maybe IncidentKey)

Identifies the incident to which this trigger event should be applied. If there's no open (i.e. unresolved) incident with this key, a new one will be created. If there's already an open incident with a matching key, this event will be appended to that incident's log. The event key provides an easy way to "de-dup" problem reports. If this field isn't provided, PagerDuty will automatically open a new incident with a unique key.

tClient :: Lens' Trigger (Maybe Text)

The name of the monitoring client that is triggering this event.

tClientUrl :: Lens' Trigger (Maybe Text)

The URL of the monitoring client that is triggering this event.

tDetails :: Lens' Trigger Object

An arbitrary JSON object containing any data you'd like included in the incident log.

Acknowledge

data Acknowledge

acknowledge :: ServiceKey -> IncidentKey -> Acknowledge

Acknowledge events cause the referenced incident to enter the acknowledged state.

While an incident is acknowledged, it won't generate any additional notifications, even if it receives new trigger events. Your monitoring tools should send PagerDuty an acknowledge event when they know someone is presently working on the problem.

aDescription :: Lens' Acknowledge (Maybe Text)

Text that will appear in the incident's log associated with this event.

aDetails :: Lens' Acknowledge Object

An arbitrary JSON object containing any data you'd like included in the incident log.

Resolve

data Resolve

resolve :: ServiceKey -> IncidentKey -> Resolve

Resolve events cause the referenced incident to enter the resolved state.

Once an incident is resolved, it won't generate any additional notifications. New trigger events with the same incident_key as a resolved incident won't re-open the incident. Instead, a new incident will be created. Your monitoring tools should send PagerDuty a resolve event when the problem that caused the initial trigger event has been fixed.

rDescription :: Lens' Resolve (Maybe Text)

Text that will appear in the incident's log associated with this event.

rDetails :: Lens' Resolve Object

An arbitrary JSON object containing any data youd like included in the incident log.

Types

class Event a

Minimal complete definition

eventType

data AnyEvent where

Constructors

Event :: Event a => a -> AnyEvent

data Response

rsStatus :: Lens' Response Text

success.

rsMessage :: Lens' Response Text

Event processed.

rsIncidentKey :: Lens' Response IncidentKey

The key of the incident that will be affected by the request.