Network.PagerDuty.REST

Sending requests

send :: (MonadIO m, FromJSON b) => SubDomain -> Auth s -> Manager -> Request a s b -> m (Either Error b)

See: sendWith

sendWith :: (MonadIO m, FromJSON b) => Env s -> Request a s b -> m (Either Error b)

paginate :: (MonadIO m, Paginate a, FromJSON b) => SubDomain -> Auth s -> Manager -> Request a s b -> Source m (Either Error b)

See: paginateWith

paginateWith :: (MonadIO m, Paginate a, FromJSON b) => Env s -> Request a s b -> Source m (Either Error b)

Configuration

Environment

data Env s

The environment containing the parameters required to make PagerDuty requests.

envDomain :: forall s. Lens' (Env s) SubDomain

envAuth :: forall s s'. Lens (Env s) (Env s') (Auth s) (Auth s')

envManager :: forall s. Lens' (Env s) Manager

envLogger :: forall s. Lens' (Env s) Logger

Logging

data Logger

The log level and associated logger function.

Constructors

None

Debug (Text -> IO ())

debug :: MonadIO m => Logger -> Text -> m ()

Log a message using the debug logger, or if none is specified noop.

Types

Requests

class Paginate a

Exists primarily to obtain a constraint for the paginate function.

data Request a s b

Errors

data Error

_Internal :: Prism' Error Text

_Integration :: Prism' Error IntegrationError

_REST :: Prism' Error RESTError

Messages

class HasMessage s a | s -> a where

Methods

message :: Lens' s a

A short human-readable message describing the error.

Fields

class HasErrors s a | s -> a where

Methods

errors :: Lens' s a

A list of human-readable reasons for the error. These values, and even their format, are subject to change.

Integration

data IntegrationError

status :: Lens' IntegrationError Text

REST

data RESTError

data Code

code :: Lens' RESTError Code

In the case of an error, the PagerDuty error code can give further details on the nature of the error.

See: description

description :: Code -> Text

Authentication

data Security

Constructors

Basic

Token

data Auth a where

Constructors

AuthBasic :: ByteString -> ByteString -> Auth Basic

AuthToken :: ByteString -> Auth Token

Keys

data Key a

mkKey :: Text -> Key a

Aliases

type ServiceKey = Key "service"

type IncidentKey = Key "incident"

Identifiers

data Id a

mkId :: Text -> Id a

Aliases

type AlertId = Id "alert"

type ContactId = Id "contact"

type EmailFilterId = Id "email-filter"

type EscalationPolicyId = Id "escalation-policy"

type EscalationRuleId = Id "escalation-rule"

type IncidentId = Id "incident"

type LogEntryId = Id "log-entry"

type NoteId = Id "note"

type NotificationRuleId = Id "notification-rule"

type OverrideId = Id "schedule-override"

type RequesterId = Id "requester"

type ScheduleId = Id "schedule"

type ServiceId = Id "service"

type UserId = Id "user"

type VendorId = Id "vendor"

type WebhookId = Id "webhook"

type WindowId = Id "maintenance-window"

SubDomain

data SubDomain

mkSubDomain :: ByteString -> SubDomain

Address

data Address

mkAddress :: Text -> Address

Miscellaneous

data Empty

Constructors

Empty