Network.PagerDuty.REST.Webhooks

Description

Webhooks let you recieve HTTP callbacks when interesting events happen within your PagerDuty account. Details surrounding the interesting event will be sent via HTTP to a URL that you specify.

PagerDuty currently supports incident-based webhooks. After adding a webhook URL to a PagerDuty service, the triggering of new incidents on that service will cause outgoing webhook messages to be sent to that URL. In addition, certain interesting changes to an incident's state will cause other types of incident webhook messages to be sent: Generally, any change to the status or assigned_to_user of an incident will cause an outgoing message to be sent.

See: http://developer.pagerduty.com/documentation/rest/webhooks

Documentation

data WebhookType

Constructors

WTrigger
Sent when an incident is newly created/triggered.

WAcknowledge
Sent when an incident has had its status changed from triggered to acknowledged.

WUnacknowledge
Sent when an incident is unacknowledged due to timeout.

WResolve
Sent when an incident has been resolved.

WAssign
Sent when an incident has been manually reassigned to another user in a different escalation chain.

WEscalate
Sent when an incident has been escalated to another user in the same escalation chain.

WDelegate
Sent when an incident has been reassigned to another escalation chain.

WOther Text
Note: As new types of incident actions are supported, new incident webhook message types will be added.

data Webhook

wId :: Lens' Webhook WebhookId

Uniquely identifies this outgoing webhook message; can be used for idempotency when processing the messages.

wType :: Lens' Webhook WebhookType

The webhook message type.

wCreatedOn :: Lens' Webhook UTCTime

The date/time when the incident changed state.