Control.Monad.Trans.PagerDuty

Transformer

type PagerDuty s = PagerDutyT s IO

A convenient alias for PagerDutyT IO.

data PagerDutyT s m a

Running

runPagerDutyT :: PagerDutyT s m a -> Env s -> m (Either Error a)

Environment

data Env s

The environment containing the parameters required to make PagerDuty requests.

envDomain :: forall s. Lens' (Env s) SubDomain

envAuth :: forall s s'. Lens (Env s) (Env s') (Auth s) (Auth s')

envManager :: forall s. Lens' (Env s) Manager

envLogger :: forall s. Lens' (Env s) Logger

scoped :: MonadReader (Env s) m => (Env s -> m a) -> m a

Logging

data Logger

The log level and associated logger function.

Constructors

None

Debug (Text -> IO ())

debug :: MonadIO m => Logger -> Text -> m ()

Log a message using the debug logger, or if none is specified noop.

Errors

data Error

Constructors

Internal Text

Integration IntegrationError

REST RESTError

_Internal :: Prism' Error Text

_Integration :: Prism' Error IntegrationError

_REST :: Prism' Error RESTError

Integration

data IntegrationError

status :: Lens' IntegrationError Text

REST

data RESTError

code :: Lens' RESTError Code

In the case of an error, the PagerDuty error code can give further details on the nature of the error.

See: description

Code

newtype Code

Constructors

Code Integer

description :: Code -> Text

Fields

message :: HasMessage s a => Lens' s a

A short human-readable message describing the error.

errors :: HasErrors s a => Lens' s a

A list of human-readable reasons for the error. These values, and even their format, are subject to change.

Lifting

hoistError :: MonadError Error m => Either Error a -> m a

Integration API

submit :: (MonadIO m, MonadReader (Env s) m, MonadError Error m) => Event -> m Response

submitCatch :: (MonadIO m, MonadReader (Env s) m) => Event -> m (Either Error Response)

REST API

send :: (MonadIO m, MonadReader (Env s) m, MonadError Error m, FromJSON b) => Request a s b -> m b

sendCatch :: (MonadIO m, MonadReader (Env s) m, FromJSON b) => Request a s b -> m (Either Error b)

paginate :: (MonadIO m, MonadReader (Env s) m, MonadError Error m, Paginate a, FromJSON b) => Request a s b -> Source m b

paginateCatch :: (MonadIO m, MonadReader (Env s) m, Paginate a, FromJSON b) => Request a s b -> Source m (Either Error b)